// src/modules/student/student-interface.ts
// import { Schema, model, connect, Model } from "mongoose";

import { Model, Types } from "mongoose";

export type TuserName = {
fristName: string;
  middleName?: string;
  lastName: string;
};
export type TGuardian = {
  fatherName: string;
  fatherOccupation: string;
  fatherContactNo: string;
  motherName: string;
  motherOccupation: string;
  motherContactNo: string;
};
export type TLocalGuardian = {
  name: string;
  occupation: string;
  contactNo: string;
  address: string;
};
export type TStudent = {
  id: string;
  user: Types.ObjectId;
  name: TuserName;
  gender: "male" | "female" | "other";
  dateOfBirth?: string;
  email: string;
  contactNo: string;
  emergencyContactNo: string;
  bloodGroup?: "A+" | "A-" | "B+" | "B-" | "AB+" | "AB-" | "O+" | "O-";
  presentAddress: string;
  permanentAddress: string;
  guardian: TGuardian;
  localGuardian: TLocalGuardian;
  profileImg?: string;
  admissionSemester : Types.ObjectId;
  isDeleted?: boolean;
};

export type StudentMethod = {
  isUserExists() : Promise< TStudent | null>;
};
export type StudentModel = Model<
  TStudent,
 Record<string, never>,
  StudentMethod
>;